% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogncac.R
\name{cognac}
\alias{cognac}
\title{cogncac: Core Gene Alignement Concatenation}
\usage{
cognac(
  fastaDir,
  featureDir,
  fastaFiles,
  featureFiles,
  fastaExt,
  featureExt,
  genomeIds,
  geneEnv,
  outDir,
  runId,
  minGeneNum,
  maxMissGenes,
  coreGeneThresh,
  copyNumTresh,
  threadVal,
  distMat,
  njTree,
  revTranslate,
  outGroup,
  partitionFile,
  subModel,
  keepTempFiles,
  percId,
  algnCovg,
  cdHitFlags
)
}
\arguments{
\item{fastaDir}{Directory containing the fasta files}

\item{featureDir}{Directory containing the Gff3 or genbank files}

\item{fastaFiles}{In place of a directory with fasta files, a character 
vector with the paths to the the fasta files for each genome can be
input. Incompatible with "fasta dir."}

\item{featureFiles}{A character vector with a gff3 or genbank files
for each genome}

\item{genomeIds}{Optional character vector with the unique ID for each 
genome. If not supplied each genome is created by removing the file 
extension from the fasta file.}

\item{geneEnv}{Optional. Environment created by "CreateGeneDataEnv", this
allows a user to parse the genomic data ahead of time. This may be 
useful when troubleshooting parameters with respect to the alignment
or cd-hit clustering step.}

\item{outDir}{Directory to write the alignments, other output files, 
and temp files created by cd-hit and mafft. Defaults to the current 
working directory.}

\item{runId}{Optional string to prepend to the output files.}

\item{minGeneNum}{An integer specifing the minimum number of genes to
use in creating the alignment. Default: 1.}

\item{maxMissGenes}{Optional double specifying what fraction of genes are
permitted to be missing to be allowed to be included in the alignmnet.
If an out-group is input these genomes are not removed.}

\item{coreGeneThresh}{Optional double specifying what fraction of genomes
that a gene must occur in to be considered core. Defaults to 0.99.}

\item{copyNumTresh}{Optional double specifying what fraction of genomes
that a gene can be duplicate in and remain in the analysis. If a gene
is duplicated, the first gene is taken and used in the alignment.
Defaults to 0.001 -- a gene can be duplicated in 1 in 1000 genomes.}

\item{threadVal}{Number of threads available. Defaults to all available 
threads.}

\item{distMat}{Optional logical to create a distance matrix. If true, a
distance matrix is calculated as the pairwise number of mutations 
between genomes. Stored in the output environment as "distMat."}

\item{njTree}{Optional logical to create a neighbor joining tree with ape.
returned in the output environment as "njTree." Defaults to false.}

\item{revTranslate}{Optional logical to to convert the aa alignment to dna.
Defaults to false.}

\item{outGroup}{Optional character vector of genomes IDs that are not used
in selecting core genes. An out-group is useful for many phylogenetic 
analysis, but the intersection of genes in all genomes in the analysis
may be smaller if phylogenetically divergent genomes are included when 
selecting core genes. This enables identification of a larger number of 
core genes while still maintaining an outgroup for down-stream analysis.}

\item{partitionFile}{Optional logical to create a partition file to input
to RAxML}

\item{subModel}{Optional string for the type of substitution model 
for partition file}

\item{keepTempFiles}{Optional logical to keep any temporary files 
generated by mafft of cd-hit.}

\item{percId}{Optional double for the identity threshold for identifying
orthologous genes by cd-hit. See "-c" in the cd-hit manual. Defaults 
to 0.7.}

\item{algnCovg}{Optional double for the maximum disparity in the lenght
of the genes to be considered orthologous by cd-hit. See -aL" in the
cd-hit manual. Defaults to 0.8.}

\item{cdHitFlags}{Optional string with parameters to pass to cd-hit to 
define clusters aside from "-c" and "-aL" that can be used to define the
  clustering parameters.}
}
\value{
An envrioment with the alignmnet data. Variables included
  by default are "aaAlgnPath" and "metaData"
}
\description{
The cognac function idetentifies shared genes to be used as phylogenetic
  markers. Genes are aligned individually wth mafft and concatenates them
  into a single alignment.
}
