% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CreateAlgnDistMat}
\alias{CreateAlgnDistMat}
\title{Create Algnment Distance Matrix}
\arguments{
\item{Path}{to the alignment}

\item{Method}{for calculating distance: "raw", or "shared."}

\item{Logical}{to specify whether to remove gap positions from the alignment
to create the core genome.}
}
\value{
A numeric matrix
}
\description{
This function reads in the path to a multiple sequence alignment. The
  pairwise alignment distances for each sequence are returned. The 
  pairwise distances are calculated in parallel via the RcppParallel 
  package.
}
