% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGeneDataEnv.R
\name{CreateGeneDataEnv}
\alias{CreateGeneDataEnv}
\title{Create a gene data environment}
\usage{
CreateGeneDataEnv(featureFiles, fastaFiles, genomeIds, outDir)
}
\arguments{
\item{featureFiles}{Character vector with the paths to the gff or genbank 
files}

\item{fastaFiles}{Character vector with the paths to the fasta files}

\item{genomeIds}{Optional vector with the unique identifiers for the 
genomes included in the analysis}

\item{outDir}{Directory in which to write the faa file with the translated
gene sequences from all genes.}
}
\value{
The environment containing a list of the genome features ("gfList"), 
  vector of gene sequences ("geneSeqs"), unique gene ids ("geneIds"), 
  path to the faa file written ("faaPath").
}
\description{
This function initializes the environment containing data on the genes
  used in the analysis. This function adds the list of parsed gff files,
  fasta file paths, and the genome Ids of the genomes included in the 
  analysis to a new environment which is returned. This function uses 
  all available threads to parse the genomic data in parallel via the 
  RcppParallel package.
}
