% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePartionFile.R
\name{CreatePartionFile}
\alias{CreatePartionFile}
\title{Create Partion File}
\usage{
CreatePartionFile(algnEnv, outDir, runId, subModel, seqType)
}
\arguments{
\item{algnEnv}{Environment created by the cognac that contains the 
alignment data.}

\item{outDir}{Directory to write the file to}

\item{runId}{Run identifier appended to the output file}

\item{subModel}{substitution model to be used}
}
\value{
void
}
\description{
This is a function creates the gene partition file used by RAxML
}
\details{
This is a function that writes the alignmnet partition file used by RAxML
to apply. Written in the format:
DNA, peg.1=1-30
DNA, peg.2=31-60
}
