% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FilterAlignmentPositions}
\alias{FilterAlignmentPositions}
\title{Filter Alignment Positions}
\arguments{
\item{msaPath}{Path to the alignment}

\item{filterMsaPath}{Path to write the filtered alignment}

\item{double}{minGapFrac=0.01,}

\item{minSubThresh}{Minimium number o subsititutions}

\item{genePositions}{Optional vector of partitions within the alignment}
}
\value{
void
}
\description{
This function
}
