% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FilterPartitionedAlgnPositions}
\alias{FilterPartitionedAlgnPositions}
\title{Filter Partitioned Algn Positions}
\arguments{
\item{msaPath}{Path to the alignment}

\item{filterMsaPath}{Path to write the filtered alignment}

\item{genePositions}{Vector of gene partitions in the alignment}

\item{minGapFrac}{Double representing the minimium fraction of gaps to
remain in the alignment.Defaults to 0.01.}

\item{minSubThresh}{Integer representing the minimum number of
substitutions to remain in the alignment. This is the number of instances
of any minor allele to remain. Defaults to 0.}
}
\value{
Updated vector with the partions in the filtered alignment
}
\description{
This function
}
