% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReverseTranslateAlgn.R
\name{ReverseTranslateAlgn}
\alias{ReverseTranslateAlgn}
\title{Reverse Translate Algnment}
\usage{
ReverseTranslateAlgn(geneEnv, concatGeneFa, outDir, runId, algnEnv)
}
\arguments{
\item{geneEnv, }{concatGeneFa, outDir}

\item{concatGeneFa}{Path to the amino acid alignment}

\item{runId}{Run identifier to append to the alignment file}

\item{algnEnv}{Environment created by congnac with the data on the alignment}
}
\value{
Path to the reverse translated alignment
}
\description{
This function reads in a concatenated gene alignment created by cognac and
}
